Create MX (#578)
* Create MX

This is a config designed for MX anticheat NO it is not short for Matrix so if someone is going to use on a matrix server they should do .config load Matrix this is a bypassing config with blatant killaura/movement checks also The template was blocksmc config I made some tweaks in rotational checks and target strafe to make sure it didn't flag and YES players can make speed as much as they want because MX is very weak at tons of movement checks like speed because It's a combat based check anticheat players should be careful if using it because they have to know that it's MX anticheat or if it's other anticheat they might get banned. There is a warning telling that MX is not short for matrix to not get confused with people and so they don't complain for a flagging config. People should use it at their own risk because if staff comes on and sees well they are going to get banned!

* Rename MX to mx.json

---------

Co-authored-by: Izuna <izuna.seikatsu@ccbluex.net>