Create grim1_8
I have made some changes to the grim config so it's made for 1.8, I have patched some things like the grim config telly scaffold kept flagging so i switched to god bridge and I enabled super knockback. The reason why I made this config is because not everyone wants to have to tune the grim config and make it work for 1.8 grim servers.