Create gladmc
This config has hypixel b hop as it uses modified version of verus which is more stricter, Anti void should be used on SHORT distances only or the anticheat will kick. Velocity somehow works and I set hypixel B hop instead of blocksmc speed because the anticheat doesn't kick with hypixel B hop while Blocksmc speed does. This config was tested in gladmc and it worked is that the problem is that If you use antivoid for way too long then you will be kicked by the anticheat. There is NO sword blocking no slow hitting through because the anticheat will kick you even in interact mode. Also it's recommended to use eagle on this config rather than scaffold because its way more stable to use eagle.